رفع اشکال

+++ ApiKarbord/Models/Public/Web_FDocH.cs
@@ -214,7 +214,7 @@
         public double? ArzRate { get; set; }
 
 
-        public int? AccSerialNumber { get; set; }
+        public long? AccSerialNumber { get; set; }
 
         public int? AccDocNo { get; set; }
     }
